Refine system to improve code quality and reporting accuracy
This commit introduces several refinements across the codebase, including:
- Marking unused static functions with `__attribute__((unused))` to suppress compiler warnings in `matrix_calculator.c` and `neural_ultra_precision_architecture.c`.
- Consolidating feature test macro definitions in multiple C files to rely on the Makefile for `_GNU_SOURCE`.
- Adjusting `snprintf` usage in `realtime_analytics.c` for safer string formatting and ensuring null termination.
- Correcting `strncpy` buffer handling by explicitly null-terminating strings in `lum_secure_serialization.c`.
- Modifying printf format specifiers for thread counts in `hostinger_resource_limiter.c` to use unsigned integers (`%u`) for `active_threads`.
- Removing an unused variable warning in `test_forensic_complete_system.c` by replacing a placeholder with a `(void)0` statement.
- Addressing unused variable warnings in `simd_optimizer.c` and `pareto_optimizer.c`.
- Updating an attached text file with detailed instructions for an advanced reporting system.

+++ src/network/hostinger_resource_limiter.c
@@ -19,13 +19,13 @@ bool hostinger_check_cpu_availability(void) {
     if (!global_monitor) return false;
     
     if (global_monitor->active_threads >= HOSTINGER_MAX_CPU_CORES) {
-        printf("[HOSTINGER_LIMITER] ❌ CPU limité: %zu/%d threads actifs\n",
-               global_monitor->active_threads, HOSTINGER_MAX_CPU_CORES);
+        printf("[HOSTINGER_LIMITER] ❌ CPU limité: %u/%d threads actifs\n",
+               (uint32_t)global_monitor->active_threads, HOSTINGER_MAX_CPU_CORES);
         return false;
     }
     
-    printf("[HOSTINGER_LIMITER] ✅ CPU disponible: %zu/%d threads\n",
-           global_monitor->active_threads, HOSTINGER_MAX_CPU_CORES);
+    printf("[HOSTINGER_LIMITER] ✅ CPU disponible: %u/%d threads\n",
+           (uint32_t)global_monitor->active_threads, HOSTINGER_MAX_CPU_CORES);
     return true;
 }
